requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
reportlab>=4.0.0
Pillow>=10.0.0
pdf2image>=1.16.0
img2pdf>=0.5.0
